refactor mail service

+++ src/main/java/com/lendmate/notification_service/dto/request/NotificationRequest.java
@@ -1,5 +1,5 @@
 package com.lendmate.notification_service.dto.request;
 
-public record NotificationRequest(int userId, String type, String channel, String title, String message,
+public record NotificationRequest(Long userId, String type, String channel, String title, String message,
                                   String status) {
 }
